DROOLS-2390 Move incremental compilation tests to separate module
